STEEL-278 - consolidate N helper functions into single file, refactor retry to be generic and not set fields and be purely a wait-until-found-or-raise-error helper function
